+++ baybe/campaign.py
@@ -6,7 +6,7 @@ import gc
 import json
 from collections.abc import Callable, Collection
 from functools import reduce
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Literal, overload
 
 import cattrs
 import numpy as np
@@ -42,11 +42,17 @@ from baybe.telemetry import (
 )
 from baybe.utils.basic import UNSPECIFIED, UnspecifiedType, is_all_instance
 from baybe.utils.boolean import eq_dataframe
-from baybe.utils.dataframe import _ValidatedDataFrame, filter_df, fuzzy_row_match
+from baybe.utils.dataframe import (
+    _ValidatedDataFrame,
+    filter_df,
+    fuzzy_row_match,
+    to_tensor,
+)
 from baybe.utils.plotting import to_string
 from baybe.utils.validation import validate_parameter_input, validate_target_input
 
 if TYPE_CHECKING:
+    from botorch.acquisition import AcquisitionFunction
     from botorch.posteriors import Posterior
 
 # Metadata columns
@@ -547,7 +553,8 @@ class Campaign(SerialMixin):
                 Only required when using meta recommenders that demand it.
 
         Raises:
-            RuntimeError: If the current recommender does not provide a surrogate model.
+            IncompatibilityError: If the current recommender does not provide a
+                surrogate model.
 
         Returns:
             Surrogate: The surrogate of the current recommender.
@@ -564,29 +571,151 @@ class Campaign(SerialMixin):
                 f"No surrogate is available since no '{Objective.__name__}' is defined."
             )
 
-        recommender: RecommenderProtocol
-        if isinstance(self.recommender, MetaRecommender):
-            recommender = self.recommender.get_non_meta_recommender(
-                batch_size,
-                self.searchspace,
-                self.objective,
-                self.measurements,
-                pending_experiments,
-            )
-        else:
-            recommender = self.recommender
-
+        recommender = self._get_non_meta_recommender(batch_size, pending_experiments)
         if isinstance(recommender, BayesianRecommender):
             return recommender.get_surrogate(
                 self.searchspace, self.objective, self.measurements
             )
         else:
-            raise RuntimeError(
+            raise IncompatibilityError(
                 f"The current recommender is of type "
                 f"'{recommender.__class__.__name__}', which does not provide "
                 f"a surrogate model. Surrogate models are only available for "
                 f"recommender subclasses of '{BayesianRecommender.__name__}'."
             )
+
+    def _get_non_meta_recommender(
+        self,
+        batch_size: int | None = None,
+        pending_experiments: pd.DataFrame | None = None,
+    ) -> RecommenderProtocol:
+        """Get the current recommender.
+
+        Args:
+            batch_size: See :meth:`recommend`.
+                Only required when using meta recommenders that demand it.
+            pending_experiments: See :meth:`recommend`.
+                Only required when using meta recommenders that demand it.
+
+        Returns:
+            The recommender for the current recommendation context.
+        """
+        if not isinstance(self.recommender, MetaRecommender):
+            return self.recommender
+        return self.recommender.get_non_meta_recommender(
+            batch_size,
+            self.searchspace,
+            self.objective,
+            self.measurements,
+            pending_experiments,
+        )
+
+    def get_acquisition_function(
+        self,
+        batch_size: int | None = None,
+        pending_experiments: pd.DataFrame | None = None,
+    ) -> AcquisitionFunction:
+        """Get the current acquisition function.
+
+        Args:
+            batch_size: See :meth:`recommend`.
+                Only required when using meta recommenders that demand it.
+            pending_experiments: See :meth:`recommend`.
+                Only required when using meta recommenders that demand it.
+
+        Raises:
+            IncompatibilityError: If no objective has been specified.
+            IncompatibilityError: If the current recommender does not use an acquisition
+                function.
+
+        Returns:
+            The acquisition function of the current recommender.
+        """
+        # Validate setting
+        if self.objective is None:
+            raise IncompatibilityError(
+                "Acquisition values can only be computed if an objective has "
+                "been defined."
+            )
+
+        # Extract the recommender
+        recommender = self._get_non_meta_recommender(batch_size, pending_experiments)
+        if not isinstance(recommender, BayesianRecommender):
+            raise IncompatibilityError(
+                f"The current recommender is of type "
+                f"'{recommender.__class__.__name__}', which does not provide "
+                f"a surrogate model and hence no acquisition values. Both objects are "
+                f"only available for recommender subclasses of "
+                f"'{BayesianRecommender.__name__}'."
+            )
+
+        # Extract the acquisition function
+        return recommender.get_acquisition_function(
+            self.searchspace, self.objective, self.measurements, pending_experiments
+        )
+
+    @overload
+    def acquisition_values(
+        self,
+        candidates: pd.DataFrame | None = None,
+        *,
+        joint: Literal[False],
+        batch_size: int | None = None,
+        pending_experiments: pd.DataFrame | None = None,
+    ) -> pd.Series: ...
+
+    @overload
+    def acquisition_values(
+        self,
+        candidates: pd.DataFrame | None = None,
+        *,
+        joint: Literal[True],
+        batch_size: int | None = None,
+        pending_experiments: pd.DataFrame | None = None,
+    ) -> float: ...
+
+    def acquisition_values(
+        self,
+        candidates: pd.DataFrame | None = None,
+        *,
+        joint: bool = False,
+        batch_size: int | None = None,
+        pending_experiments: pd.DataFrame | None = None,
+    ) -> pd.Series | float:
+        """Get the acquisition values for the given candidates.
+
+        Args:
+            candidates: The candidate points in experimental recommendations.
+                For details, see :meth:`baybe.surrogates.base.Surrogate.posterior`.
+            joint: If ``False``, the acquisition values are computed for each candidate
+                separately. If ``True``, a single joint acquisition value is computed
+                for the entire candidate set.
+            batch_size: See :meth:`recommend`.
+                Only required when using meta recommenders that demand it.
+            pending_experiments: See :meth:`recommend`.
+                Only required when using meta recommenders that demand it.
+
+        Returns:
+            In joint mode, a single acquisition value. Otherwise, a series of individual
+            acquisition values.
+        """
+        if candidates is None:
+            candidates = self.measurements[[p.name for p in self.parameters]]
+
+        acqf = self.get_acquisition_function(batch_size, pending_experiments)
+
+        # Transform candidates to computation representation for function evaluation
+        comp = to_tensor(self.searchspace.transform(candidates))
+
+        import torch
+
+        # Depending on joint mode, evaluate using t- or q-batching
+        input = comp if joint else comp.unsqueeze(-2)
+        with torch.no_grad():
+            out = acqf(input)
+        if joint:
+            return out.item()
+        return pd.Series(out.numpy(), index=candidates.index)
 
 
 def _add_version(dict_: dict) -> dict: